Feat : page change feed and home
